move session oriented methods to application controller
- delete helpers/session_helper.rb
- declare logged_in? and current_user as a helper described in controllers/applicaation_controller.rb
- fix wrong usage of current_user(@current_user -> current_user)
