+++ eumetopias_2/app/controllers/application_controller.rb
@@ -1,7 +1,26 @@
 class ApplicationController < ActionController::Base
-  include SessionsHelper
   rescue_from ActionController::RoutingError, with: :render_404
   rescue_from StandardError, with: :render_500
+  helper_method :current_user, :logged_in?
+
+  def log_in(user)
+    session[:user_id] = user.id
+  end
+
+  def current_user
+    if session[:user_id]
+      @current_user ||= User.find_by(id: session[:user_id])
+    end
+  end
+
+  def logged_in?
+    !current_user.nil?
+  end
+
+  def log_out
+    session.delete(:user_id)
+    @current_user = nil
+  end
 
   def render_404
     logger.debug('404')
