+++ eumetopias_2/app/views/layouts/_header.html.erb
@@ -2,7 +2,7 @@
   <nav class="navbar navbar-dark bg-dark">
     <%= link_to "Eumetopias", '/', id: "logo", class: "navbar-brand mb-0 h1" %>
     <% if logged_in? %>
-      <span class="navbar-text"><%= @current_user.name %></span>
+      <span class="navbar-text"><%= current_user.name %></span>
       <%= link_to t('dictionary.button.logout'), '/logout', method: :delete, class: "nav-link" %>
     <% else %>
         <%= link_to t('dictionary.button.login'), '/login', class: "nav-link" %>
